refactor: STOMP PreSend 및 PostSend 핸들러 제거 및 관련 테스트 삭제
- ConnectPreSendHandler, ConnectPostSendHandler, CustomStompChannelInterceptor 제거로 불필요한 핸들러 정리
- 관련된 테스트 코드 삭제
- 핸들러 기반 로직을 Spring WebSocket 이벤트 기반으로 이전 완료
- 코드 가독성 및 유지보수성 개선
