+++ backend/src/main/java/coffeeshout/global/websocket/event/SessionSubscribeEventListener.java
@@ -24,7 +24,7 @@ public class SessionSubscribeEventListener {
         final StompHeaderAccessor accessor = StompHeaderAccessor.wrap(event.getMessage());
         final String destination = accessor.getDestination();
         final String subscriptionId = accessor.getSubscriptionId();
-        
+
         // 플레이어 정보 가져오기 (있으면)
         String playerInfo = "";
         if (sessionManager.hasPlayerKey(sessionId)) {
@@ -33,13 +33,13 @@ public class SessionSubscribeEventListener {
             final String playerName = sessionManager.extractPlayerName(playerKey);
             playerInfo = String.format(", joinCode=%s, playerName=%s", joinCode, playerName);
         }
-        
-        log.info("구독 시작: sessionId={}, destination={}, subscriptionId={}{}", 
+
+        log.info("구독 시작: sessionId={}, destination={}, subscriptionId={}{}",
                 sessionId, destination, subscriptionId, playerInfo);
-        
+
         // 구독 정보 추가
         subscriptionInfoService.addSubscription(sessionId, destination, subscriptionId);
-        
+
         // INFO 레벨에서도 상세 구독 정보 로깅 (구독자 수 포함)
         subscriptionInfoService.logSubscriptionInfo(destination);
     }
@@ -49,19 +49,16 @@ public class SessionSubscribeEventListener {
         final String sessionId = event.getMessage().getHeaders().get("simpSessionId", String.class);
         final StompHeaderAccessor accessor = StompHeaderAccessor.wrap(event.getMessage());
         final String subscriptionId = accessor.getSubscriptionId();
-        
-        // 플레이어 정보 가져오기 (있으면)
-        String playerInfo = "";
-        if (sessionManager.hasPlayerKey(sessionId)) {
-            final String playerKey = sessionManager.getPlayerKey(sessionId);
-            final String joinCode = sessionManager.extractJoinCode(playerKey);
-            final String playerName = sessionManager.extractPlayerName(playerKey);
-            playerInfo = String.format(", joinCode=%s, playerName=%s", joinCode, playerName);
-        }
-        
-        log.info("구독 해제: sessionId={}, subscriptionId={}{}", 
-                sessionId, subscriptionId, playerInfo);
-        
+
+//        // 플레이어 정보 가져오기 (있으면)
+//        String playerInfo = "";
+//        if (sessionManager.hasPlayerKey(sessionId)) {
+//            final String playerKey = sessionManager.getPlayerKey(sessionId);
+//            final String joinCode = sessionManager.extractJoinCode(playerKey);
+//            final String playerName = sessionManager.extractPlayerName(playerKey);
+//            playerInfo = String.format(", joinCode=%s, playerName=%s", joinCode, playerName);
+//        }
+//
         // 구독 정보 제거 - subscriptionId로 정확한 destination 찾아서 제거
         subscriptionInfoService.removeSubscriptionById(sessionId, subscriptionId);
     }
